requirements.txt
flask==3.0.3
flask-limiter==3.8.0
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
groq==0.9.0